requirements.txt
altair==4.2.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs==22.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.5
cachetools==5.3.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.0.1
click==8.1.3
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/b/abs_1419earm7u/croot/comm_1671231131638/work
contourpy==1.0.7
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
entrypoints @ file:///C:/ci/entrypoints_1649926621128/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
graphviz==0.20.1
idna==3.4
importlib-metadata==6.0.0
ipykernel @ file:///C:/b/abs_b4f07tbsyd/croot/ipykernel_1672767104060/work
ipython @ file:///C:/b/abs_536xsn7rj0/croot/ipython_1670919360443/work
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client @ file:///C:/b/abs_b9pns5mx5p/croot/jupyter_client_1671703062216/work
jupyter_core @ file:///C:/b/abs_84df679bho/croot/jupyter_core_1672332237650/work
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mdurl==0.1.2
mlxtend==0.21.0
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
numpy==1.24.1
opencv-python==4.7.0.68
packaging @ file:///C:/b/abs_cfsup8ur87/croot/packaging_1671697442297/work
pandas==1.5.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.20.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==10.0.1
pydeck==0.8.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305.1
pyzmq @ file:///C:/ci/pyzmq_1657615952984/work
requests==2.28.2
rich==13.2.0
scikit-learn==1.2.1
scipy==1.10.0
semver==2.13.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
smmap==5.0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.17.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado @ file:///C:/ci/tornado_1662458743919/work
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
wincertstore==0.2
zipp==3.11.0
